Add RFC7523 compliant OAuth access token endpoint
This endpoint handles the exchange of a JWT authorization grant to an
API access token as described in RFC7523 [1].

Partners will generate JWT tokens for their users and sign them with
their client secret. They then pass that JWT token as a configuration to
the Hypothesis client (that part still needs work), and it will then
exchange the JWT authorization grant for an access token via the `POST
/api/token` endpoint.

RFC7523 requires the use of the RS256 algorithm to sign the JWT token,
the implementation in this commit still uses thi HS256 algorithm. In the
future, authclients would generate a public/private keypair and share
the public part with us.

[1]: https://tools.ietf.org/html/rfc7523